+++ src/api/sip.ts
@@ -164,7 +164,6 @@ export class SIP extends EventEmitter {
         incoming: ev.rtcSession.direction === 'incoming',
         partyNumber,
         partyName,
-        remoteVideoEnabled: ev.remoteWithVideo,
         localVideoEnabled: ev.withVideo,
         sessionStatus: ev.sessionStatus,
         callConfig: getCallConfigFromHeader(m?.getHeader('X-WEBPHONE-CALL')),
@@ -238,19 +237,28 @@ export class SIP extends EventEmitter {
       this.emit('session-updated', {
         id: ev.sessionId,
         videoSessionId: ev.videoClientSessionId,
-        remoteVideoEnabled: true,
         remoteVideoStreamObject: videoSession.remoteStreamObject,
+        localStreamObject: session.localVideoStreamObject,
+        videoClientSessionTable: Object.entries(
+          session.videoClientSessionTable,
+        ).map(([key, value]) => ({ ...value, vId: key })),
+        remoteUserOptionsTable: session.remoteUserOptionsTable,
       })
     })
     phone.addEventListener('videoClientSessionEnded', ev => {
       if (!ev) {
         return
       }
+      const session = phone.getSession(ev.sessionId)
       this.emit('session-updated', {
         id: ev.sessionId,
         videoSessionId: ev.videoClientSessionId,
-        remoteVideoEnabled: false,
         remoteVideoStreamObject: null,
+        videoClientSessionTable: Object.entries(
+          session.videoClientSessionTable,
+        ).map(([key, value]) => ({ ...value, vId: key })),
+        localStreamObject: session.localVideoStreamObject,
+        remoteUserOptionsTable: session.remoteUserOptionsTable,
       })
     })
 
@@ -281,11 +289,11 @@ export class SIP extends EventEmitter {
       //   this.enableVideo(ev.sessionId)
       // }
 
-      // TODO DuyP to fix issues related to video call
-      // this.emit('session-updated', {
-      //   id: ev.sessionId,
-      //   remoteUserOptionsTable: ev.remoteUserOptionsTable,
-      // })
+      // DuyP add this to handle toggle on/off video streams
+      this.emit('session-updated', {
+        id: ev.sessionId,
+        remoteUserOptionsTable: ev.remoteUserOptionsTable,
+      })
     })
 
     phone.addEventListener('rtcErrorOccurred', ev => {
@@ -413,6 +421,10 @@ export class SIP extends EventEmitter {
     this.phone?.setWithVideo(sessionId, false)
   setMuted = (muted: boolean, sessionId: string) =>
     this.phone?.setMuted({ main: muted }, sessionId)
+  setMutedVideo = (muted: boolean, sessionId: string) => {
+    this.phone?.setMuted({ videoClient: muted }, sessionId)
+  }
+
   switchCamera = async (sessionId: string, isFrontCamera: boolean) => {
     // alert(this.currentFrontCamera)
     if (!this.phone) {
@@ -426,9 +438,13 @@ export class SIP extends EventEmitter {
     if (this.cameraIds === undefined || this.cameraIds.length === 0) {
       return
     }
-    const cameras = this.cameraIds.map(s => s.deviceId)
-    this.currentCamera = isFrontCamera ? cameras[1] : cameras[0]
 
+    const cameras = this.cameraIds.map(s => s.deviceId)
+    if (cameras.length < 2) {
+      return
+    }
+
+    this.currentCamera = isFrontCamera ? cameras[0] : cameras[1]
     const videoOptions = {
       call: {
         mediaConstraints: sipCreateMediaConstraints(
@@ -442,7 +458,11 @@ export class SIP extends EventEmitter {
           isFrontCamera,
         ),
       },
+      shareStream: true,
     }
+
+    // TODO: Need handle the best way to switch camera still keep connection
+    // current: disable video then enable, same with UC desktop
     this.phone?.setWithVideo(sessionId, false, videoOptions)
     this.phone?.setWithVideo(sessionId, true, videoOptions)
   }
@@ -543,6 +563,7 @@ const getWebrtcClient = (dtmfSendPal = false, sourceId?: string) =>
         answer: {
           mediaConstraints: sipCreateMediaConstraints(sourceId, true),
         },
+        shareStream: true,
       },
     },
     dtmfSendPal,