+++ src/brekekejs/index.d.ts
@@ -385,7 +385,10 @@ export type Sip = {
   makeCall: MakeCallFn
   answer: MakeCallFn
   setWithVideo(sessionId: string, withVideo?: boolean): void
-  setMuted(options: { main: boolean }, sessionId: string): void
+  setMuted(
+    options: { main?: boolean; videoClient?: boolean },
+    sessionId: string,
+  ): void
   setWithVideo(
     sessionId: string,
     withVideo?: boolean,
@@ -438,6 +441,7 @@ export type SipConstructorOptions = {
       answer: {
         mediaConstraints: MediaStreamConstraints
       }
+      shareStream?: boolean
     }
   }
   configuration?: Partial<SipConfiguration>
@@ -473,6 +477,7 @@ export type SipEventMap = {
   videoClientSessionCreated: VideoSession
   videoClientSessionEnded: VideoSession
   rtcErrorOccurred: Error
+  remoteUserOptionsChanged: Session
 }
 export type PhoneStatusChangedEvent = {
   phoneStatus: 'starting' | 'started' | 'stopping' | 'stopped'
@@ -501,6 +506,7 @@ export type Session = {
   remoteWithVideo: boolean
   remoteStreamObject: MediaStream
   localStreamObject: MediaStream
+  localVideoStreamObject: MediaStream
   incomingMessage?: {
     getHeader(h: string): string | undefined
     body?: object
@@ -521,6 +527,7 @@ export type Session = {
   }
   remoteUserOptionsTable: null
   analyzer: null
+  user: string
 }
 export type VideoSession = {
   sessionId: string