+++ src/components/VideoPlayer.web.tsx
@@ -1,5 +1,5 @@
 import { observer } from 'mobx-react'
-import { ActivityIndicator, StyleSheet } from 'react-native'
+import { ActivityIndicator, StyleSheet, View } from 'react-native'
 
 const css = StyleSheet.create({
   loading: {
@@ -10,22 +10,26 @@ const css = StyleSheet.create({
   },
 })
 
-export const VideoPlayer = observer((p: { sourceObject: MediaStream }) =>
-  p.sourceObject ? (
-    <video
-      ref={video => {
-        if (video) {
-          video.srcObject = p.sourceObject
-        }
-      }}
-      style={{
-        width: '100%',
-        height: '100%',
-        objectFit: 'cover',
-      }}
-      autoPlay
-    />
-  ) : (
-    <ActivityIndicator style={css.loading} />
-  ),
+export const VideoPlayer = observer(
+  (p: { sourceObject: MediaStream; isShowLoading?: boolean }) =>
+    p.sourceObject ? (
+      <video
+        ref={video => {
+          if (video) {
+            video.srcObject = p.sourceObject
+          }
+        }}
+        style={{
+          width: '100%',
+          height: '100%',
+          objectFit: 'cover',
+        }}
+        playsInline
+        autoPlay
+      />
+    ) : p.isShowLoading ? (
+      <ActivityIndicator style={css.loading} />
+    ) : (
+      <View style={css.loading} />
+    ),
 )